% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createLikertData.R
\name{createLikertData}
\alias{createLikertData}
\title{createLikertData}
\usage{
createLikertData(likert_scale_questions, participant_factors, question_factors,
  response_levels)
}
\arguments{
\item{likert_scale_questions}{The data consisting of likert-scale items with the
response_levels supplied below. The columns should be either factorial or numeric,
containing only whole numbers (e.g. 0 - 5, 1 - 5, ... )}

\item{participant_factors}{Factors that can be used to classify the participants}

\item{question_factors}{Factors that can be used to classify the questions}

\item{response_levels}{The answer choices for the likert questions.
Often c( "strongly agree", "agree", ... )}
}
\description{
Create data of class likertData from three dataframes. This function also
applies some sanity checks for the inputs.
}
\author{
Gregor de Cillia
}

