% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createTestData.R
\name{createTestData}
\alias{createTestData}
\title{createTestData}
\usage{
createTestData( participants = 100,
                questions    = 10,
                participant_factors = list(
                  country  = c( "US", "AT", "CH" ),
                  gender   = c( "Male", "Female" ),
                  position = c( "Programmer",
                                "Project leader",
                                "Salesman" ) ),
                question_factors = list(
                  difficulty = c( "easy", "hard" ),
                  skill      = c( "logic",
                                  "compassion",
                                  "leadership" ) ),
                response_levels = c( "Strongly disagree",
                                     "Disagree",
                                     "Neutral",
                                     "Agree",
                                     "Strongly Agree" ))
}
\arguments{
\item{participants}{either the number of participants or a vector of
participant names}

\item{questions}{either the number of questions or a vector of
question names}

\item{participant_factors}{list containing all participant factors as names
and their possible levels as entries.}

\item{question_factors}{list containing all question factors}

\item{response_levels}{character vector with response levels}
}
\description{
Create a test data set based on uniform distributions. For checking
statistical test using p values, this data set can confirm wether
the p-values are equally distributed if this data is used.
}
\examples{
td = createTestData()
td
summary( td )

td = createTestData(
  c( "John", "Tom", "Anna", "Gustavo" ),
  c( "The new changes in my company improved my motivation to work",
     "Our IT system needs improvement" )
)

# overwrite the randomly generated genders
td$row_factors$gender = factor( c( "Male", "Male", "Female", "Male" ) )
summary( td )
td$row_factors

}

