% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likertData.R
\name{plot.likertData}
\alias{plot.likertData}
\title{Plot likert data}
\usage{
\method{plot}{likertData}(x, ...)
}
\arguments{
\item{x}{An object of class likertData}

\item{...}{further arguments passed down to the HH::likert call}
}
\description{
Create a simple likert plot from the data
}
\examples{
testData = createTestData()
plot( testData )
plot( testData, positive.order = TRUE )
}
\author{
Gregor de Cillia
}

