% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyLikertOutput.R
\name{plot.shinyLikertOutput}
\alias{plot.shinyLikertOutput}
\title{plot shiny likert output}
\usage{
\method{plot}{shinyLikertOutput}(x, ...)
}
\arguments{
\item{x}{object to be printed}

\item{...}{currently unused}
}
\description{
creates the app in a fluidpage.
}

