% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likertData.R
\name{print.likertData}
\alias{print.likertData}
\title{print likert data}
\usage{
\method{print}{likertData}(x, ...)
}
\arguments{
\item{x}{An object of class likertData}

\item{...}{further arguments. Will be ignored}
}
\description{
prints some information about the data
}
\examples{
testData2 = createTestData( 100, 10 )
print( testData2 )
}
\author{
Gregor de Cillia
}

