% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renderFactorOverview.R
\name{renderFactorOverview}
\alias{renderFactorOverview}
\title{Create widgets to visualize the factor variables}
\usage{
renderFactorOverview(data, id = toString(paste0("id", sample(1:10000, 1))),
  input = data$input, output = data$output, env = parent.frame(),
  quoted = FALSE, expr = "NULL", ...)
}
\arguments{
\item{data}{the datasetto be shown}

\item{id}{a unique id of the output}

\item{input}{argument for shiny session}

\item{output}{argument for shiny session}

\item{env}{argument for shiny session}

\item{quoted}{argument for shiny session}

\item{expr}{argument for shiny session}
}
\description{
Create widgets to visualize the factor variables
}

