% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renderShinyLikert.R
\name{renderShinyLikert}
\alias{renderShinyLikert}
\title{Shiny likert}
\usage{
renderShinyLikert( id, data,
                   input, output,
                   dropdown_factors = NULL,
                   questions = names( data$likert_data ),
                   height = NULL,
                   env = parent.frame(),
                   quoted = FALSE,
                   expr= "NULL",
                   response_levels = levels( data$likert_data[,1] ),
                   split_factors = NULL,
                   ...
                 )
}
\arguments{
\item{id}{A string definig the id of the output. The id has to be
unique within your project}

\item{data}{An object of type likertData to be plotted}

\item{input, output}{options to be passed down from the following call \cr
shinyApp( ui=..., server = function( input,output, session )\{ ... \} )}

\item{dropdown_factors}{names ot the factors which should be used as for
the dropdown menus}

\item{questions}{the questions that should be displayed. (Currently broken)}

\item{height}{minimum and maximum height of the plot in px. Setting this option
will create a slider to control the height of the plot.}

\item{env}{argument to be used by the shiny runtime}

\item{quoted}{argument to be used by the shiny runtime}

\item{expr}{argument to be used by the shiny runtime}

\item{response_levels}{answer possibilities of interest}

\item{split_factors}{factors to be used in the HH plot}

\item{...}{further arguments to be passed down to HH::likert}
}
\value{
A list of rendered shiny objects which can be used as outputs.
}
\description{
Creates a list of rendered shiny outputs to be used inside a call to
shinyApp() or inside an ShinyMarkdown document
}
\examples{
x <- 1
# shinyApp(
#     ui= fluidPage(
#         uiOutput( "selector" ),
#         plotOutput( "plot" )
#     ),
#     server = function( input, output, session ){
#         rendered = renderShinyLikert(
#             "ID 1", testData2 ,
#             input, output,
#             dropdown = "gender" )
#         output$plot     = rendered$plot
#         output$selector = rendered$selector
#     }
# )
}
\author{
Gregor de Cillia
}

