% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likertData.R
\name{summary.likertData}
\alias{summary.likertData}
\title{Summarize likert data}
\usage{
\method{summary}{likertData}(object, ...)
}
\arguments{
\item{object}{An object of class likertData}

\item{...}{further arguments for the HH::likert call}
}
\description{
prints a summary about the data
}
\examples{
testData2 = createTestData()
summary( testData2 )
}
\author{
Gregor de Cillia
}

