% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createTestData.R
\name{createTestData}
\alias{createTestData}
\title{createTestData}
\usage{
createTestData(participants, questions, input, output,
  participant_factors = list(country = c("US", "AT", "CH"), gender = c("Male",
  "Female"), position = c("Programmer", "Project leader", "Salesman")),
  question_factors = list(difficulty = c("easy", "hard"), skill = c("logic",
  "compassion", "leadership")), response_levels = c("Strongly disagree",
  "Disagree", "Neutral", "Agree", "Strongly Agree"))
}
\arguments{
\item{participants}{Either the number of participants or a vector of participant names}

\item{questions}{Either the number of questions or a vector of question names}

\item{participant_factors}{A list containing all participant factors}

\item{question_factors}{A list containing all question factors}

\item{response_levels}{Vector with response levels}
}
\description{
Create a test data set based on uniform distributions
}
\examples{
createTestData( 100, 10 )

td = createTestData( c( "John", "Tom", "Anna", "Gustavo" ),
                     c( "The new changes in my company improved my motivation to work",
                        "Our IT system needs improvement"
                         )  )
td$row_factors$gender = factor( c( "Male", "Male", "Female", "Male" ) )
summary( td )
td$row_factors

}

