% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createTestData.R
\name{createTestData}
\alias{createTestData}
\title{createTestData}
\usage{
createTestData( participants, questions,
                       input, output,
                       participant_factors, question_factors,
                       response_levels )
}
\arguments{
\item{participants}{either the number of participants or a vector of participant names}

\item{questions}{either the number of questions or a vector of question names}

\item{input}{arguments for the shiny session}

\item{output}{arguments for the shiny session}

\item{participant_factors}{a list containing all participant factors}

\item{question_factors}{a list containing all question factors}

\item{response_levels}{Vector with response levels}
}
\description{
Create a test data set based on uniform distributions. For checking
statistical test using p values, this data set can confirm wether
the p-values are equally distributet for data generated by uniform
distributions.
}
\examples{
createTestData( 100, 10 )

td = createTestData( c( "John", "Tom", "Anna", "Gustavo" ),
                     c( "The new changes in my company improved my motivation to work",
                        "Our IT system needs improvement"
                         )  )
td$row_factors$gender = factor( c( "Male", "Male", "Female", "Male" ) )
summary( td )
td$row_factors

}

