% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renderShinyLikert.R
\name{renderShinyLikert}
\alias{renderShinyLikert}
\title{Shiny likert}
\usage{
renderShinyLikert( data,
                   dropdown_factors = NULL,
                   questions = names( data$likert_data ),
                   height = NULL,
                   response_levels = levels( data$likert_data[,1] ),
                   split_factors = NULL,
                   group,
                   grouping,
                   id,
                   ...
                 )
}
\arguments{
\item{data}{An object of type likertData to be plotted}

\item{dropdown_factors}{names ot the factors which should be used as for
the dropdown menus}

\item{questions}{the questions that should be displayed. (Currently broken)}

\item{height}{minimum and maximum height of the plot in px. Setting this option
will create a slider to control the height of the plot.}

\item{response_levels}{answer possibilities of interest}

\item{split_factors}{factors to be used in the HH plot}

\item{group}{variables to group the plot}

\item{grouping}{chooses wether HH or likert will be used to display
the result}

\item{id}{A string definig the id of the output. The id has to be
unique within your project}

\item{...}{further arguments to be passed down to HH::likert}
}
\value{
A list of rendered shiny objects which can be used as outputs.
}
\description{
Creates a list of rendered shiny outputs to be used inside a call to
shinyApp() or inside an ShinyMarkdown document
}
\examples{
\dontrun{
  shinyApp(
      ui= fluidPage(
          uiOutput( "selector" ),
          uiOutput( "plot" )
      ),
      server = function( input, output, session ){
          testData2 = createTestData()
          rendered = renderShinyLikert(
              testData2 ,
              dropdown = "gender" )
          output$plot     = rendered$plot
          output$selector = rendered$selector
      }
  )
}
}
\author{
Gregor de Cillia
}

