% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renderShinyLikert.R
\name{renderShinyLikert}
\alias{renderShinyLikert}
\title{Shiny likert}
\usage{
renderShinyLikert( data,
                   dropdown_factors = NULL,
                   questions = names( data$likert_data ),
                   height = NULL,
                   response_levels = levels( data$likert_data[ , 1 ] ),
                   split_factors = NULL,
                   group = NULL,
                   grouping = "likert",
                   id = toString( sample( 10^15, 1 ) ),
                   wrap = 30,
                   custom_tests = NULL,
                   ...
                 )
}
\arguments{
\item{data}{an object of class likertData to be plotted}

\item{dropdown_factors}{a character vector containing names of the factors which
should be used for the dropdown menus}

\item{questions}{the questions that should be displayed. (Currently broken)}

\item{height}{minimum and maximum height of the plot in px. Setting this
option will create a slider to control the height of the
plot.}

\item{response_levels}{answer possibilities to be plotted and tested}

\item{split_factors}{factors to be used in the HH plot}

\item{group}{variables to group the plot}

\item{grouping}{decides wether the package HH or the package likert
will be used to display the result. Only relevant if
group has been set}

\item{id}{a string definig the id of the output. The id has to be
unique within your project}

\item{wrap}{apply wraping for text on the y-Axis (factor levels).}

\item{custom_tests}{a list containing tests to be applied in the table
output}

\item{...}{further arguments to be passed down to HH::likert}
}
\value{
A list of rendered shiny objects which can be used as outputs.
}
\description{
Creates a list of rendered shiny outputs to be used in a shiny app or
a "shiny-markdown" document
}
\examples{
\dontrun{
  library( shiny )
  shinyApp(
      ui= fluidPage(
          uiOutput( "selector" ),
          uiOutput( "plot" )
      ),
      server = function( input, output, session ){
          testData = createTestData()
          rendered = renderShinyLikert(
              testData,
              dropdown = "gender" )
          output$plot     = rendered$plot
          output$selector = rendered$selector
      }
  ) }
}
\author{
Gregor de Cillia
}

